Refactor Error class search to its own class
To avoid duplication the error class method has been moved to Error
Parse class
