+++ lib/easy_meli.rb
@@ -3,6 +3,7 @@ require 'httparty'
 require 'easy_meli/version'
 require 'easy_meli/constants'
 require 'easy_meli/errors'
+require 'easy_meli/error_parser'
 require 'easy_meli/configuration'
 require 'easy_meli/authorization_client'
 require 'easy_meli/api_client'
